+++ packages/vite-plugin-super-template/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.2
+
+- Upgrade to `@visma/react-intl-bundled-messages@^1.0.0`
+
 # 0.1.1
 
 - Set dev server port to 3000